requirements.txt
docopt==0.6.2
python-dateutil==2.6.1
six==1.10.0
requests==2.18.4
